* fix, network.h: aligned size of tcp_flags field to effective usage.

+++ src/network.h
@@ -439,7 +439,7 @@ struct pkt_data {
   pm_counter_t pkt_num;
   pm_counter_t flo_num;
   u_int8_t flow_type;
-  u_int32_t tcp_flags; /* XXX */
+  u_int8_t tcp_flags;
   struct timeval time_start;
   struct timeval time_end;
 };
